fastapi/citation_system/requirements.txt
setuptools>=65.0.0
wheel
torch==2.3.1
transformers==4.36.0
colbert-ir==0.2.14
faiss-cpu==1.12.0
numpy>=1.24.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
tqdm==4.66.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
jieba==0.42.1
